wsconn: exit receiver and sender loops when stream ends
Instead of looping forever and matching on `.next()`, we can use `while let` to
loop for as long as the stream will provide data in a more idiomatic manner.

This fixes a bug in the receiver loop where the `None` branch does nothing,
looping over and over again and pegging one core per connection that's broken.

The sender loop does not suffer from this but it does make the matching code less indented and more straightforward.